requirements.txt
annotated-types==0.6.0
anyio==4.6.2
asttokens==2.0.5
attrs==24.2.0
Bottleneck==1.3.7
click==8.1.7
comm==0.2.1
debugpy==1.6.7
decorator==5.1.1
executing==0.8.3
fastapi==0.112.2
fastjsonschema==2.16.2
greenlet==3.0.1
h11==0.14.0
idna==3.7
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.7.2
matplotlib-inline==0.1.6
nbformat==5.10.4
nest-asyncio==1.6.0
numexpr==2.8.7
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pandasql==0.7.3
parso==0.8.3
pexpect==4.8.0
pip==24.2
platformdirs==3.10.0
prompt-toolkit==3.0.43
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.15.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==25.1.2
referencing==0.30.2
rpds-py==0.10.6
setuptools==75.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.34
stack-data==0.2.0
starlette==0.38.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2023.3
uvicorn==0.20.0
wcwidth==0.2.5
wheel==0.44.0
sklearn-pandas==2.2.0

